WIP: cleaning up by comment
