+++ ai_chatbot_backend/.env.example
@@ -26,6 +26,9 @@ admin_password=123
 # Directory path for file storage (relative to project root)
 DATA_DIR=data
 
+# Database configuration
+DATABASE_URL="sqlite:///./courses.db"
+
 # LLM Configuration  
 # Options: local, remote, mock
 # local = use local model (requires GPU and model files)
