+++ ai_chatbot_backend/.envrc
@@ -38,12 +38,6 @@ if [[ -n "$VIRTUAL_ENV" && -d "$VIRTUAL_ENV" ]]; then
     export ENVIRONMENT="dev"
     export DEBUG="true"
     
-    # Database configuration
-    export DATABASE_URL="sqlite:///./courses.db"
-    # Alternative database paths:
-    # export DATABASE_URL="sqlite:///./data/my_database.db"
-    # export DATABASE_URL="sqlite:////absolute/path/to/database.db"
-
     echo "🚀 Activated ai_chatbot_backend environment (Poetry)"
     echo "📍 Virtual env: $VIRTUAL_ENV"
     echo "🐍 Python: $(python --version 2>&1)"
