+++ ai_chatbot_backend/app/api/router.py
@@ -5,7 +5,6 @@ from app.api.routes import (
     file_completions,
     file_tester,
     files,
-    file_metadata,
     problems,
 )
 from fastapi import APIRouter
@@ -34,9 +33,6 @@ api_router.include_router(courses.router, prefix="/courses", tags=["courses"])
 
 # Files management
 api_router.include_router(files.router, prefix="/files", tags=["files"])
-
-# File metadata management
-api_router.include_router(file_metadata.router, prefix="/file-metadata", tags=["file-metadata"])
 
 # Problems management
 api_router.include_router(problems.router, prefix="/problems", tags=["problems"])
